requirements.txt
# DEV.to Analytics Platform - Requirements
# Modern async PostgreSQL-first architecture

# Core async framework
httpx>=0.27.0          # Async HTTP client (replaces requests)
asyncpg>=0.29.0        # Fast async PostgreSQL driver

# FastAPI and Web Framework
fastapi>=0.128.0       # Modern async web framework
uvicorn[standard]>=0.40.0  # ASGI server with WebSocket and extras
pydantic>=2.12.0       # Data validation and serialization

# Database
SQLAlchemy[asyncio]>=2.0.0  # Async ORM and Core
psycopg2-binary>=2.9.9       # PostgreSQL adapter (for sync operations)
pgvector>=0.2.0              # Vector similarity search

# NLP and Text Analysis
spacy>=3.7.0           # NLP models and entity recognition
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl  # spaCy English model
vaderSentiment>=3.3.2  # Sentiment analysis
beautifulsoup4>=4.12.0 # HTML parsing
lxml>=5.0.0            # Fast XML/HTML parser

# CLI and scripting
python-dotenv>=1.0.0    # Environment variable management
tqdm>=4.66.0            # Progress bars
rich>=13.7.0            # Beautiful terminal output (for migration script)
python-dateutil>=2.8.2  # Advanced date parsing

# Data processing (for NLP services)
numpy>=1.24.0           # Required for pgvector and spaCy
pandas>=2.0.0           # Data analysis (optional, for future analytics)

# Old synchronous libraries (for legacy scripts compatibility)
requests>=2.31.0        # HTTP client (used by old scripts)

# Optional: Development and testing
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
